game should rules the board and the pawn
