+++ src/test/java/com/marmelab/quoridor/game/GameTest.java
@@ -1,9 +1,13 @@
 package com.marmelab.quoridor.game;
 
 import com.marmelab.quoridor.model.Board;
+import com.marmelab.quoridor.model.Fence;
 import com.marmelab.quoridor.model.Pawn;
+import com.marmelab.quoridor.model.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -19,6 +23,220 @@ public class GameTest {
         Game game = new Game(board);
         //Then
         assertThat(game.getPawn()).isEqualTo(expected);
+    }
+
+    @Test
+    @DisplayName("by default, no fences are on the board")
+    void constructorFences() {
+        // Given
+        Board board = new Board(3);
+        Game game = new Game(board);
+        // When
+        List<Fence> fences = game.getFences();
+        // Then
+        assertThat(fences).isEmpty();
+    }
+
+    @Test
+    @DisplayName("add the fence should add the fence at the right place")
+    void addFenceHappyPath() {
+        // Given
+        Board board = new Board(3);
+        Game game = new Game(board);
+        Fence expected = new Fence(0, 0, false);
+        // When
+        game.addFence(new Fence(0, 0, false));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsOnly(expected);
+    }
+
+    @Test
+    @DisplayName("it is not possible to add another fence at the same place")
+    void addFenceNotAthTheSamePlace() {
+        // Given
+        Board board = new Board(3);
+        Game game = new Game(board);
+        Fence expected = new Fence(0, 0, false);
+        game.addFence(new Fence(0, 0, false));
+        // When
+        game.addFence(new Fence(0, 0, true));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsOnly(expected);
+    }
+
+    @Test
+    @DisplayName("get fences should return the right number of vertical fences")
+    void getFencesReturnsTheRightNumberOfVerticalFences() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, false));
+        game.addFence(new Fence(0, 2, false));
+        // When
+        List<Fence> fences = game.getFences();
+        // Then
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(new Position(0, 0), false),
+                new Fence(new Position(0, 2), false));
+    }
+
+    @Test
+    @DisplayName("get fences should return the right number of horizontal fences")
+    void getFencesReturnsTheRightNumberOfHorizontalFences() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, true));
+        game.addFence(new Fence(2, 0, true));
+        // When
+        List<Fence> fences = game.getFences();
+        // Then
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 0, true),
+                new Fence(2, 0, true));
+    }
+
+    @Test
+    @DisplayName("It should not be possible to add an horizontal fence one square after a fence")
+    void addFenceNotPossibleHorizontalFenceNextSquare() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, true));
+        game.addFence(new Fence(1, 0, true));
+        // When
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 0, true));
+    }
+
+    @Test
+    @DisplayName("It should not be possible to add a vertical fence one square after a fence")
+    void addFenceNotPossibleVerticalFenceNextFence() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, false));
+        game.addFence(new Fence(0, 1, false));
+        // When
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 0, false));
+    }
+
+    @Test
+    @DisplayName("It should not be possible to add a vertical fence one horizontal fence")
+    void addFenceNotPossibleVerticalFenceOnHorizontalFence() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, true));
+        // When
+        game.addFence(new Fence(0, 0, false));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 0, true));
+    }
+
+    @Test
+    @DisplayName("It should not be possible to add an horizontal fence one vertical fence")
+    void addFenceNotPossibleHorizontalFenceOnVerticalFence() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, false));
+        // When
+        game.addFence(new Fence(0, 0, true));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 0, false));
+    }
+
+    @Test
+    @DisplayName("It should not be possible to add a vertical fence before a vertical fence")
+    void addFenceNotPossibleVerticalFenceBeforeVerticalFence() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 1, false));
+        // When
+        game.addFence(new Fence(0, 0, false));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 1, false));
+    }
+
+    @Test
+    @DisplayName("It should not be possible to add an horizontal fence before a horizontal fence")
+    void addFenceNotPossibleHorizontalFenceBeforeHorizontalFence() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(1, 0, true));
+        // When
+        game.addFence(new Fence(0, 0, true));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(1, 0, true));
+    }
+
+    @Test
+    @DisplayName("It should be possible to add a vertical fence between two horizontal fences")
+    void addFenceBetweenTwoHorizontalFences() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, true));
+        game.addFence(new Fence(2, 0, true));
+        // When
+        game.addFence(new Fence(1, 0, false));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 0, true),
+                new Fence(2, 0, true),
+                new Fence(1, 0, false));
+    }
+
+    @Test
+    @DisplayName("It should not be possible to add an horizontal fence between two vertical fences")
+    void addFenceBetweenTwoVerticalFences() {
+        // Given
+        Board board = new Board(5);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, false));
+        game.addFence(new Fence(0, 2, false));
+        // When
+        game.addFence(new Fence(0, 1, true));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 0, false),
+                new Fence(0, 2, false),
+                new Fence(0, 1, true));
+    }
+
+    @Test
+    @DisplayName("It should be possible to add a fence which closes the access to the goal line")
+    void addFenceNotCrossable() {
+        // Given
+        Board board = new Board(3);
+        Game game = new Game(board);
+        game.addFence(new Fence(0, 0, false));
+        // When
+        game.addFence(new Fence(0, 1, true));
+        // Then
+        List<Fence> fences = game.getFences();
+        assertThat(fences).containsExactlyInAnyOrder(
+                new Fence(0, 0, false));
     }
 
 }
